backend/requirements.txt
Flask>=2.0.1
Flask-SQLAlchemy>=2.5.1
Flask-Bcrypt>=0.7.1
Flask-CORS>=3.0.10
PyJWT>=2.3.0
stegano>=0.9.9
cryptography>=3.4.7
pytest>=7.3.1